Factor out symlinkPackages and implement LinkCommand with it.
