+++ src/commands/BootstrapCommand.js
@@ -1,10 +1,9 @@
-import FileSystemUtilities from "../FileSystemUtilities";
 import NpmUtilities from "../NpmUtilities";
 import PackageUtilities from "../PackageUtilities";
+import LinkUtilities from "../LinkUtilities";
 import Command from "../Command";
 import async from "async";
 import find from "lodash.find";
-import path from "path";
 
 export default class BootstrapCommand extends Command {
   initialize(callback) {
@@ -89,36 +88,6 @@ export default class BootstrapCommand extends Command {
   }
 
   /**
-   * Create a symlink to a dependency's binary in the node_modules/.bin folder
-   * @param {String} src
-   * @param {String} dest
-   * @param {String} name
-   * @param {String|Object} bin
-   * @param {Function} callback
-   */
-  createBinaryLink(src, dest, name, bin, callback) {
-    const destBinFolder = path.join(dest, ".bin");
-    // The `bin` in a package.json may be either a string or an object.
-    // Normalize to an object.
-    const bins = typeof bin === "string"
-      ? { [name]: bin }
-      : bin;
-    const srcBinFiles = [];
-    const destBinFiles = [];
-    Object.keys(bins).forEach((name) => {
-      srcBinFiles.push(path.join(src, bins[name]));
-      destBinFiles.push(path.join(destBinFolder, name));
-    });
-    // make sure when have a destination folder (node_modules/.bin)
-    const actions = [(cb) => FileSystemUtilities.mkdirp(destBinFolder, cb)];
-    // symlink each binary
-    srcBinFiles.forEach((binFile, idx) => {
-      actions.push((cb) => FileSystemUtilities.symlink(binFile, destBinFiles[idx], "exec", cb));
-    });
-    async.series(actions, callback);
-  }
-
-  /**
    * Install external dependencies for all packages
    * @param {Function} callback
    */
@@ -156,79 +125,6 @@ export default class BootstrapCommand extends Command {
    * @param {Function} callback
    */
   symlinkPackages(callback) {
-    this.logger.info("Symlinking packages and binaries");
-    this.progressBar.init(this.filteredPackages.length);
-    const actions = [];
-    this.filteredPackages.forEach((filteredPackage) => {
-      // actions to run for this package
-      const packageActions = [];
-      Object.keys(filteredPackage.allDependencies)
-        // filter out external dependencies and incompatible packages
-        .filter((dependency) => {
-          const match = this.packageGraph.get(dependency);
-          return match && filteredPackage.hasMatchingDependency(match.package);
-        })
-        .forEach((dependency) => {
-          // get Package of dependency
-          const dependencyPackage = this.packageGraph.get(dependency).package;
-          // get path to dependency and its scope
-          const { location: dependencyLocation } = dependencyPackage;
-          const dependencyPackageJsonLocation = path.join(dependencyLocation, "package.json");
-          // ignore dependencies without a package.json file
-          if (!FileSystemUtilities.existsSync(dependencyPackageJsonLocation)) {
-            this.logger.error(
-              `Unable to find package.json for ${dependency} dependency of ${filteredPackage.name},  ` +
-              "Skipping..."
-            );
-          } else {
-            // get the destination directory name of the dependency
-            const pkgDependencyLocation = path.join(filteredPackage.nodeModulesLocation, dependencyPackage.name);
-            // check if dependency is already installed
-            if (FileSystemUtilities.existsSync(pkgDependencyLocation)) {
-              const isDepSymlink = FileSystemUtilities.isSymlink(pkgDependencyLocation);
-              // installed dependency is a symlink pointing to a different location
-              if (isDepSymlink !== false && isDepSymlink !== dependencyLocation) {
-                this.logger.warn(
-                  `Symlink already exists for ${dependency} dependency of ${filteredPackage.name}, ` +
-                  "but links to different location. Replacing with updated symlink..."
-                );
-              // installed dependency is not a symlink
-              } else if (isDepSymlink === false) {
-                this.logger.warn(
-                  `${dependency} is already installed for ${filteredPackage.name}. ` +
-                  "Replacing with symlink..."
-                );
-                // remove installed dependency
-                packageActions.push((cb) => FileSystemUtilities.rimraf(pkgDependencyLocation, cb));
-              }
-            }
-            // ensure destination path
-            packageActions.push((cb) => FileSystemUtilities.mkdirp(
-              pkgDependencyLocation.split(path.sep).slice(0, -1).join(path.sep), cb
-            ));
-            // create package symlink
-            packageActions.push((cb) => FileSystemUtilities.symlink(
-              dependencyLocation, pkgDependencyLocation, "junction", cb
-            ));
-            const dependencyPackageJson = require(dependencyPackageJsonLocation);
-            if (dependencyPackageJson.bin) {
-              const destFolder = filteredPackage.nodeModulesLocation;
-              packageActions.push((cb) => {
-                this.createBinaryLink(dependencyLocation, destFolder, dependency, dependencyPackageJson.bin, cb);
-              });
-            }
-          }
-        });
-      actions.push((cb) => {
-        async.series(packageActions, (err) => {
-          this.progressBar.tick(filteredPackage.name);
-          cb(err);
-        });
-      });
-    });
-    async.series(actions, (err) => {
-      this.progressBar.terminate();
-      callback(err);
-    });
+    LinkUtilities.symlinkPackages(this.filteredPackages, this.packageGraph, this.progressBar, this.logger, callback);
   }
 }
